requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via langchain
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.4.0
    # via
    #   -r requirements.in
    #   httpx
    #   openai
attrs==24.2.0
    # via aiohttp
certifi==2024.7.4
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via
    #   -r requirements.in
    #   requests
colorama==0.4.6
    # via -r requirements.in
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via
    #   -r requirements.in
    #   openai
dnspython==2.6.1
    # via
    #   -r requirements.in
    #   pymongo
exceptiongroup==1.2.1
    # via -r requirements.in
filelock==3.15.4
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fonttools==4.54.1
    # via matplotlib
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
futures==3.0.5
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.67.0
    # via
    #   grpcio-tools
    #   qdrant-client
grpcio-tools==1.67.0
    # via qdrant-client
h11==0.14.0
    # via
    #   -r requirements.in
    #   httpcore
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.5
    # via
    #   -r requirements.in
    #   httpx
httpx[http2]==0.27.0
    # via
    #   -r requirements.in
    #   openai
    #   qdrant-client
huggingface-hub==0.23.4
    # via
    #   -r requirements.in
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==6.0.1
    # via h2
idna==3.7
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
    #   requests
    #   yarl
intel-openmp==2021.4.0
    # via
    #   -r requirements.in
    #   mkl
jinja2==3.1.4
    # via
    #   -r requirements.in
    #   torch
joblib==1.4.2
    # via
    #   -r requirements.in
    #   pynndescent
    #   scikit-learn
jsonpatch==1.33
    # via
    #   -r requirements.in
    #   langchain-core
jsonpointer==3.0.0
    # via
    #   -r requirements.in
    #   jsonpatch
kiwisolver==1.4.7
    # via matplotlib
langchain==0.2.7
    # via -r requirements.in
langchain-core==0.2.12
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.2.2
    # via
    #   -r requirements.in
    #   langchain
langdetect==1.0.9
    # via -r requirements.in
langsmith==0.1.84
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
llvmlite==0.43.0
    # via
    #   -r requirements.in
    #   numba
    #   pynndescent
lxml==5.2.2
    # via
    #   -r requirements.in
    #   python-docx
markupsafe==2.1.5
    # via
    #   -r requirements.in
    #   jinja2
matplotlib==3.9.2
    # via -r requirements.in
mkl==2021.4.0
    # via -r requirements.in
mpmath==1.3.0
    # via
    #   -r requirements.in
    #   sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.3
    # via
    #   -r requirements.in
    #   torch
numba==0.60.0
    # via
    #   -r requirements.in
    #   pynndescent
    #   umap-learn
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   langchain
    #   matplotlib
    #   numba
    #   qdrant-client
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
    #   umap-learn
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.77
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
openai==1.35.10
    # via -r requirements.in
orjson==3.10.6
    # via
    #   -r requirements.in
    #   langsmith
packaging==24.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   langchain-core
    #   matplotlib
    #   pytesseract
    #   transformers
pillow==10.4.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pytesseract
    #   sentence-transformers
portalocker==2.10.1
    # via qdrant-client
propcache==0.2.0
    # via yarl
protobuf==5.28.3
    # via grpcio-tools
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   qdrant-client
pydantic-core==2.20.1
    # via
    #   -r requirements.in
    #   pydantic
pymongo==4.8.0
    # via -r requirements.in
pynndescent==0.5.13
    # via
    #   -r requirements.in
    #   umap-learn
pyparsing==3.2.0
    # via matplotlib
pypdfium2==4.30.0
    # via -r requirements.in
pytesseract==0.3.10
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via matplotlib
python-docx==1.1.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   langchain
    #   langchain-core
    #   transformers
qdrant-client==1.12.0
    # via -r requirements.in
regex==2024.5.15
    # via
    #   -r requirements.in
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   tiktoken
    #   transformers
safetensors==0.4.3
    # via
    #   -r requirements.in
    #   transformers
scikit-learn==1.5.1
    # via
    #   -r requirements.in
    #   pynndescent
    #   sentence-transformers
    #   umap-learn
scipy==1.14.0
    # via
    #   -r requirements.in
    #   pynndescent
    #   scikit-learn
    #   sentence-transformers
    #   umap-learn
sentence-transformers==3.0.1
    # via -r requirements.in
six==1.16.0
    # via
    #   -r requirements.in
    #   langdetect
    #   python-dateutil
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.36
    # via langchain
sympy==1.13.0
    # via
    #   -r requirements.in
    #   torch
tbb==2021.13.0
    # via
    #   -r requirements.in
    #   mkl
tenacity==8.5.0
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-core
threadpoolctl==3.5.0
    # via
    #   -r requirements.in
    #   scikit-learn
tiktoken==0.7.0
    # via -r requirements.in
tokenizers==0.19.1
    # via
    #   -r requirements.in
    #   transformers
torch==2.4.1
    # via sentence-transformers
tqdm==4.66.4
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
    #   umap-learn
transformers==4.42.3
    # via
    #   -r requirements.in
    #   sentence-transformers
triton==3.0.0
    # via torch
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   sqlalchemy
    #   torch
umap==0.1.1
    # via -r requirements.in
umap-learn==0.5.6
    # via -r requirements.in
urllib3==2.2.2
    # via
    #   -r requirements.in
    #   qdrant-client
    #   requests
validators==0.31.0
    # via -r requirements.in
wheel==0.43.0
    # via -r requirements.in
yarl==1.16.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
